+++ src/solverpy/solver/plugins/status/tptp.py
@@ -1,3 +1,4 @@
+from typing import Any
 import re
 
 from ..decorator import Decorator
@@ -10,10 +11,21 @@ class Tptp(Decorator):
    def __init__(self, **kwargs):
       Decorator.__init__(self, **kwargs)
 
-   def decorate(self, cmd, instance, strategy):
+   def decorate(
+      self, 
+      cmd : str, 
+      instance: Any, 
+      strategy: Any
+   ) -> str:
       return cmd
 
-   def update(self, instance, strategy, output, result):
+   def update(
+      self, 
+      instance: Any, 
+      strategy: Any, 
+      output: str, 
+      result: dict
+   ) -> None:
       status = patterns.single(TPTP_STATUS, output, None)
       if status:
          result["status"] = status